+++ .github/workflows/ci.yml
@@ -17,5 +17,7 @@ jobs:
           npx playwright install-deps
       - name: Lint (JavaScript)
         run: npm run lint
+      - name: Unit Tests (Browser)
+        run: npm run test:browser
       - name: Unit Tests (Server)
         run: npm run test:server
