+++ src/browser/fixture.js
