+++ src/browser/focus.js
