+++ src/browser/reset.js
