--- src/index.js
@@ -1,3 +1,0 @@
-export { aTimeout, defineCE, expect, html, nextFrame, oneEvent, waitUntil } from '@open-wc/testing';
-export { createConfig, getBrowsers } from './server/wtr-config.js';
-export { fixture } from './fixture/fixture.js';
