+++ src/server/index.js
@@ -1,0 +1,1 @@
+export { createConfig, getBrowsers } from './wtr-config.js';
